+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support generating identity proofs with `signify` tool.
 - Include all unicode emojis in results when searching for a shortcode in emoji picker.
+- Focus on search input when emoji picker is opened.
 
 ### Changed